backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "uvicorn",
    "openai>=1.0.0",
    "python-dotenv",
    "tiktoken",
    "pymongo[srv]>=4.15.3",
]
